requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
SQLAlchemy==1.4.46
psycopg2==2.9.5
psycopg2-binary==2.9.5
alembic==1.9.2
wily==1.24.0
bandit==1.7.4
python-dotenv==0.21.1
python-multipart==0.0.5
fastapi-security==0.5.0
python-jose==3.3.0
passlib==1.7.4